debug: recursive PML4 access not working
* ref extern kernel_pml4 from c
* drop identity mapping early in boot
* use higher-half mb2_info address in kernel
* allow terminal to take a NULL framebuffer
* fix map_page PTE_PRESENT bitmask check
* simplify physical memory allocator by allocating from last available page (temp)

+++ src/include/kernel/mb2.h
@@ -41,6 +41,6 @@ struct mb2_info {
     uint32_t reserved;
 } __attribute__((packed));
 
-struct mb2_tag *mb2_find_tag(uint32_t mb2_info, uint32_t type);
+struct mb2_tag *mb2_find_tag(uint64_t mb2_info, uint32_t type);
 
 #endif /* __ARGIR__MB2_H */
